+++ mpv.net/Misc/Help.cs
@@ -7,7 +7,7 @@ class ConsoleHelp
     public static void WriteError(object obj)
     {
         Console.ForegroundColor = ConsoleColor.Red;
-        Console.WriteLine(obj);
+        Console.WriteLine("[mpvnet] " + obj);
         Console.ResetColor();
         Trace.WriteLine(obj);
     }