+++ Changelog.md
@@ -4,7 +4,9 @@
 - new: menu item 'View > Show Progress' (p key) to show progress bar
 - new: `script-message mpv.net playlist-first`, unlike mpv does not
        restart if the first file is already active
-
+- new: if mpv.net is started from the terminal and an error happens
+       then the error is printed to the terminal instead of shown
+       with a message box
 - fix: update routine did only work when mpv.net was located in 'Program Files'
 - fix: errors were just ignored and only seen printed in the terminal in case mpv.net
        was started from the terminal, now for every error a message box is shown